ca-api-adk-streaming/pyproject.toml
[project]
name = "ca-api-agent"
version = "0.1.0"
description = "Conversational Analytics Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv==1.0.1",
    "google-adk==1.15.1",
    "immutabledict==4.2.1",
    "google-cloud-aiplatform[adk,agent-engines]==1.118.0",
    "absl-py==2.2.2",
    "pydantic==2.11.3",
    "google-auth==2.40.3",
    "google-auth-oauthlib==1.2.2",
    "google-auth-httplib2==0.2.0",
    "google-genai==1.39.1",
    "typing-extensions==4.15.0",
    "google-cloud-geminidataanalytics==0.10.0",
]


[project.optional-dependencies]
dev = [
    "google-cloud-aiplatform[adk,agent-engines,evaluation]==1.118.0",
    "pytest==8.3.5",
    "pytest-asyncio==0.26.0",
    "google-adk[eval]==1.15.1",
]


[build-system]
requires = ["uv_build>=0.7.19,<0.8.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "ca_api_agent"
module-root = ""

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[tool.pytest.ini_options]
# This section is for pytest configuration and does not need to be changed.
console_output_style = "progress"
addopts = "-vv -s"
#addopts = "-vv -s --pdb"
testpaths = ["tests/"]
log_level = "DEBUG"
log_cli = true
log_auto_indent = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
filterwarnings = [
    #"error",
    "ignore::UserWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    #'ignore:function ham\(\) is deprecated:DeprecationWarning',
]
